Move onboarding screens into their own namespace
